requirements.txt
requests~=2.24.0
fire~=0.3.1
ping3~=2.6.5
dnspython~=1.16.0
simplejson~=3.17.0
rich~=3.0.0